Refs #20 rename to nscsim
