+++ nscsim/__init__.py
@@ -1,10 +1,10 @@
 # flake8: noqa: F401
 
-"""Top-level package for nsc."""
+"""Top-level package for nscsim."""
 from __future__ import absolute_import
 
 __author__ = 'Jose Borreguero'
 __email__ = 'borreguero@gmail.com'
 __version__ = '0.0.0'
 
-from nsc import qvec, elastic
+from nscsim import qvec, elastic
