+++ nscsim/elastic/__init__.py
@@ -1,4 +1,4 @@
 # flake8: noqa: F401
 from __future__ import absolute_import
 
-from nsc.elastic import structure_factor
+from nscsim.elastic import structure_factor
