+++ nscsim/elastic/structure_factor.py
